+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # Changelog
 
+## [Unreleased]
+### Improvements
+ - Logging command is run in background
+
 ## [0.0.1]
-- Initial release
+ - Initial release
+